requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.7.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
cloud-sql-python-connector[asyncpg]>=1.4.4
google-cloud-aiplatform>=1.38.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
langgraph>=0.2.60
